updated movie component
